Add an attribute map to the User class
Various commands and other code may want to keep implementation-specific
data in the User objects. Adding this data directly to the User class as
fields and accessor methods would cause quick escalation of the class'es
code size, reduce readability, and - most importantly - be unextensible as
hell. A better solution is to add a type-safe attribute map, similar to
that of Netty's Channels. This way anyone can keep their private data on a
User object without modifying the User class.