Move the IRCd class to src/test
The IRCd class serves only for testing the library, and its functionality
will be implemented by whatever project will be using the library,
so there's no point keeping it in src/main and distributing it in the
main jar.

If we later have a dedicated place for examples of library usage, we can
put it there, but right now src/test must suffice.