Improve error handling in JOIN and PART implementation
Public IRCCommands' methods that take arguments other than (user, args)
can be called by others to make the provided user do something
(force them to join/part/quit/etc.).

Therefore, these methods should not report errors to the subject user, but
to the caller. If the caller acts on behalf of the user, he'll report the
error by sending an appropriate numeric response. If the caller acts on
behalf of someone else, then they should be able to report the errors to
whoever they work for.